Create pdf2img
here i just feed a pdf that i need in txt format in and get the pdf in image, so pytesseract has an easier time processing it in images one by one, since feeding the whole pdf to pytesseract explodes the ram.  so I found that if I break the PDF into images pytesseract can eat it in bite sized chunks, it is reasonably fast for what it does.