Create install-dependencies
I think these are all the dependencies required for the rest of the code.  I haven't double checked, so if a module isn't found error pops up you probably just have to add that module to these dependencies.  sorry if so, I don't have time to check, but i think this is the complete dependencies to run a pytesseract extraction sesh.