Create pdf2img
I found that pytesseract will blow out the ram if you feed it a whole pdf to process, so my solution was to break up the pdf into a folder of images, and then feed that folder to pytesseract for bite sized chunks for it to process.  so you feed a pdf to this code block, and it creates a folder of images of each page in the pdf for pytesseract.