Create rename-files-123
pdf2img names the files all crazy and pytesseract will process in a totally random order.  so need to rename all files with just the last 3 numbers.  note that if you run this code for a pdf with more than 999 pages you will need to adjust the code to 1234.  

this just renames the files in the folder so pytesseract processes them in the right order.