Create feed-the-beast
you'll want to point the input folder to the folder of png images and let pytesseract eat them and digest them and poop out the text file.

pytesseract has a few configurations you might play around with.  that is oem and psm.  i found that psm 6 worked good for treating it as a uniform codeblock.  other psm's will make it behave differently as it processes the data.  

oem is what OCR pytesseract is using.  oem 3 seems to be the best for what I was doing.  oem 2 makes it glitch out, I'm still learning pytesseract, but for the purposes of this tool oem 3 is fine